refactor test and fix stub
